[parquet-sdk][token_v2] migrate token ownership table pt1 ]
