+++ rust/processor/src/db/common/models/token_v2_models/mod.rs
@@ -3,3 +3,5 @@ pub mod raw_v1_token_royalty;
 pub mod raw_v2_token_activities;
 pub mod raw_v2_token_datas;
 pub mod raw_v2_token_metadata;
+pub mod raw_v2_token_ownerships;
+pub mod v2_token_utils;
