+++ rust/processor/src/db/postgres/models/fungible_asset_models/v2_fungible_asset_utils.rs
@@ -5,8 +5,9 @@
 #![allow(clippy::extra_unused_lifetimes)]
 
 use crate::{
-    db::postgres::models::{
-        token_models::token_utils::URI_LENGTH, token_v2_models::v2_token_utils::ResourceReference,
+    db::{
+        common::models::token_v2_models::v2_token_utils::ResourceReference,
+        postgres::models::token_models::token_utils::URI_LENGTH,
     },
     utils::util::{deserialize_from_string, truncate_str, Aggregator},
 };
