+++ rust/processor/src/db/postgres/models/token_v2_models/mod.rs
@@ -7,4 +7,3 @@ pub mod v2_token_activities;
 pub mod v2_token_datas;
 pub mod v2_token_metadata;
 pub mod v2_token_ownerships;
-pub mod v2_token_utils;
